+++ rattler433/rootfs/rattler/device_creator/main.py
@@ -56,6 +56,10 @@ def create_devices() -> None:
             msgs.extend(create_temp_f(manu, model, channel, id, name))
         elif device["type"] == "temp_c":
             msgs.extend(create_temp_c(manu, model, channel, id, name))
+        elif device["type"] == "temp_f_to_c":
+            msgs.extend(create_temp_f_to_c(manu, model, channel, id, name))
+        elif device["type"] == "temp_c_to_f":
+            msgs.extend(create_temp_c_to_f(manu, model, channel, id, name))
         elif device["type"] == "temp_hum_f":
             msgs.extend(create_temp_hum_f(manu, model, channel, id, name))
         elif device["type"] == "temp_hum_c":
@@ -375,6 +379,48 @@ def create_temp_c(manu: str, model: str, channel: str, id: str, nm: str) -> List
 
     # Create temp:
     payload = _create_temp_c(manufacturer=manu, model=model, dev_name=nm, uid=uid)
+    topic = f"{disc}/sensor/{_mstr(manu)}_{_mstr(model)}_{_mstr(uid)}/temp/config"
+    msgs.append((topic, payload, 2, True))
+
+    # Create time:
+    payload = _create_time(manufacturer=manu, model=model, dev_name=nm, uid=uid)
+    topic = f"{disc}/sensor/{_mstr(manu)}_{_mstr(model)}_{_mstr(uid)}/time/config"
+    msgs.append((topic, payload, 2, True))
+
+
+def create_temp_f_to_c(manu: str, model: str, channel: str, id: str, nm: str) -> List[Tuple]:
+
+    uid = channel + "/" + id
+    msgs = []
+    # Create battery:
+    payload = _create_battery(manufacturer=manu, model=model, dev_name=nm, uid=uid)
+    topic = f"{disc}/sensor/{_mstr(manu)}_{_mstr(model)}_{_mstr(uid)}/battery/config"
+    msgs.append((topic, payload, 2, True))
+
+    # Create temp:
+    payload = _create_temp_f_to_c(
+        manufacturer=manu, model=model, dev_name=nm, channel=channel, uid=uid
+    )
+    topic = f"{disc}/sensor/{_mstr(manu)}_{_mstr(model)}_{_mstr(uid)}/temp/config"
+    msgs.append((topic, payload, 2, True))
+
+    # Create time:
+    payload = _create_time(manufacturer=manu, model=model, dev_name=nm, uid=uid)
+    topic = f"{disc}/sensor/{_mstr(manu)}_{_mstr(model)}_{_mstr(uid)}/time/config"
+    msgs.append((topic, payload, 2, True))
+
+
+def create_temp_c_to_f(manu: str, model: str, channel: str, id: str, nm: str) -> List[Tuple]:
+
+    uid = channel + "/" + id
+    msgs = []
+    # Create battery:
+    payload = _create_battery(manufacturer=manu, model=model, dev_name=nm, uid=uid)
+    topic = f"{disc}/sensor/{_mstr(manu)}_{_mstr(model)}_{_mstr(uid)}/battery/config"
+    msgs.append((topic, payload, 2, True))
+
+    # Create temp:
+    payload = _create_temp_c_to_f(manufacturer=manu, model=model, dev_name=nm, uid=uid)
     topic = f"{disc}/sensor/{_mstr(manu)}_{_mstr(model)}_{_mstr(uid)}/temp/config"
     msgs.append((topic, payload, 2, True))