+++ rattler433/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## What’s changed
 
-## 🎉 New Feature
+## 🐛 Bug fixes
 
-- Add temperature only sensor, for devices such as an Acurite 986 fridge/freezer sensor.
+- Add temp_c_to_f and temp_f_to_c sensors that were left out of last update, for temp only sensors.